Bug Fix: Bug caused by parseInt
The previous behaviour is that the input is evaluated to a correct answer when the integer response is appended to a 0 or any non-numeric character. This is due to parseInt's nature. By maintaining the string type of the input and converting the solution of the arithmetic question into a string, we can ensure that only the correct integer answer will evaluate into a correct answer.